[JAX] Fix test failure due to upcoming change to JAX.
An upcoming change to JAX adds a `@jit` decorator around a number of array operators, in this case division (`/`). A side effect of that change is that large Python integer constants that overflow an int32 or int64 type may produce an error. The workaround is either to explicitly cast the large constants to a specific type (e.g., `np.float64`), or in this case we can just do the math in question in classic NumPy since it is computing test expectations.

PiperOrigin-RevId: 390111122